Fixes lint
